LayoutTests/media/modern-media-controls/icon-button/icon-button.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>IconButton</code> class.");

const layoutDelegate = {};

const iconButton = new IconButton({ layoutDelegate });

shouldBeEqualToString("iconButton.element.localName", "button");
shouldBeEqualToString("iconButton.element.className", "icon");

debug("");
debug("There should be no iconName by default");
shouldBeEqualToString("iconButton.iconName", "");

debug("");
debug("Setting an iconName with invalid layout traits should throw");
shouldThrow(function() { iconButton.iconName = Icons.Pause });
shouldBeEqualToString("iconButton.iconName", "");

debug("");
debug("Setting macOS inline layout traits and setting play icon");
layoutDelegate.layoutTraits = LayoutTraits.macOS;
iconButton.iconName = Icons.Pause;

let numberOfFrames = 0;

scheduler.frameDidFire = function()
{
    numberOfFrames++;

    if (numberOfFrames === 1) {
        shouldBeTrue('iconButton.element.style.webkitMaskImage.includes("macOS/pause")');
        shouldBeEqualToString("iconButton.element.style.width", "22px");
        shouldBeEqualToString("iconButton.element.style.height", "24px");
        shouldBeEqualToString("iconButton.element.style.webkitMaskSize", "22px 24px");

        shouldBeEqualToString("iconButton.element.getAttribute('aria-label')", `${Icons.Pause.label}`);
        debug("");
        debug("Setting start icon");
        iconButton.iconName = Icons.Start;
    } else if (numberOfFrames === 2) {
        shouldBeTrue('iconButton.element.style.webkitMaskImage.includes("macOS/start")');
        shouldBeEqualToString("iconButton.element.style.width", "70px");
        shouldBeEqualToString("iconButton.element.style.height", "70px");
        shouldBeEqualToString("iconButton.element.style.webkitMaskSize", "70px 70px");
        shouldBeEqualToString("iconButton.element.getAttribute('aria-label')", `${Icons.Start.label}`);

        finishMediaControlsTest();
    } 
};

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
