LayoutTests/media/modern-media-controls/icon-button/icon-button-expected.txt

Testing the IconButton class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS iconButton.element.localName is "button"
PASS iconButton.element.className is "icon"

There should be no iconName by default
PASS iconButton.iconName is ""

Setting an iconName with invalid layout traits should throw
PASS function () { iconButton.iconName = Icons.Pause } threw exception Could not identify icon's platform from layout traits..
PASS iconButton.iconName is ""

Setting macOS inline layout traits and setting play icon
PASS iconButton.element.style.webkitMaskImage.includes("macOS/pause") is true
PASS iconButton.element.style.width is "22px"
PASS iconButton.element.style.height is "24px"
PASS iconButton.element.style.webkitMaskSize is "22px 24px"
PASS iconButton.element.getAttribute('aria-label') is "Pause"

Setting start icon
PASS iconButton.element.style.webkitMaskImage.includes("macOS/start") is true
PASS iconButton.element.style.width is "70px"
PASS iconButton.element.style.height is "70px"
PASS iconButton.element.style.webkitMaskSize is "70px 70px"
PASS iconButton.element.getAttribute('aria-label') is "Start"
PASS successfullyParsed is true

TEST COMPLETE
